requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
dnspython==1.16.0
requests
python-dotenv
gTTS
opuslib
asyncio
youtube-dl
discord.py[voice] ~= 1.4.0
pip
ffmpeg
google-cloud-firestore